uv/tests: update error message that suggests a fork
This is an interesting update that previously used a message without a
fork occurring, but now notes a split. The dependencies in this case
are:

    dependencies = [
      '''fork-marker-disjoint-a>=2; sys_platform == "linux"''',
      '''fork-marker-disjoint-a<2; sys_platform == "linux"''',
    ]

In this case, the markers are the same for conflicting dependency
specifications. So the "no resolution" result is correct. But
previously, we wouldn't fork here, because we wouldn't consider the
marker expressions to be disjoint.

But with overlapping markers, we are actually forking here. Namely,
we consider a fork with `sys_platform == "linux"` and another with
`sys_platform != "linux"`.

I actually don't think this is even related to overlapping markers. The
previous behavior seems not quite right (even though it led to the same
result), because it implies we weren't really considering the "linux"
and "not linux" cases as distinct in resolution.