uv/tests: update preferences test
From looking at this carefully, I believe the result here is still
correct but "different" from the status quo based on the order in which
forks are processed. (At time of writing, we do sort forks, but we don't
use comprehensive criteria. So at least in some cases, the order depends
on the order that the forks were created. Since the overlapping markers
code completely changes how we generate forks, that seems plausible
here.)